// src/app/country/interfaces/country.interface.ts
export enum Region{
    Africa='Africa',
    America='Americas',
    Asia='Asia',
    Europe='Europe',
    Oceania='Oceania'
}

export interface SmallCountry{
    name:string;
    cca3:string;
    borders:string[];
}

export interface Country {
    name:         Name;
    tld?:         string[];
    cca2:         string;
    ccn3?:        string;
    cca3:         string;
    cioc?:        string;
    independent?: boolean;
    status:       Status;
    unMember:     boolean;
    currencies:   Currencies;
    idd:          Idd;
    capital:      string[];
    altSpellings: string[];
    region:       string;
    subregion:    string;
    languages:    { [key: string]: string };
    translations: { [key: string]: Translation };
    latlng:       number[];
    landlocked:   boolean;
    area:         number;
    demonyms:     Demonyms;
    flag:         string;
    maps:         Maps;
    population:   number;
    gini?:        { [key: string]: number };
    fifa?:        string;
    car:          Car;
    timezones:    string[];
    continents:   string[];
    flags:        Flags;
    coatOfArms:   CoatOfArms;
    startOfWeek:  StartOfWeek;
    capitalInfo:  CapitalInfo;
    postalCode?:  PostalCode;
    borders?:     string[];
}

export interface CapitalInfo {
    latlng: number[];
}

export interface Car {
    signs: string[];
    side:  Side;
}

export enum Side {
    Left = "left",
    Right = "right",
}

export interface CoatOfArms {
    png?: string;
    svg?: string;
}


export interface Currencies {
    EUR?: TartuGecko;
    RSD?: TartuGecko;
    DKK?: TartuGecko;
    FOK?: TartuGecko;
    ALL?: TartuGecko;
    GBP?: TartuGecko;
    GIP?: TartuGecko;
    SEK?: TartuGecko;
    ISK?: TartuGecko;
    CHF?: TartuGecko;
    PLN?: TartuGecko;
    JEP?: TartuGecko;
    RON?: TartuGecko;
    NOK?: TartuGecko;
    BYN?: TartuGecko;
    GGP?: TartuGecko;
    CZK?: TartuGecko;
    BAM?: BAM;
    BGN?: TartuGecko;
    MDL?: TartuGecko;
    IMP?: TartuGecko;
    HUF?: TartuGecko;
    MKD?: TartuGecko;
    UAH?: TartuGecko;
    RUB?: TartuGecko;
}

export interface TartuGecko {
    name:   string;
    symbol: string;
}

export interface BAM {
    name: string;
}

export interface Demonyms {
    eng:  Eng;
    fra?: Eng;
}

export interface Eng {
    f: string;
    m: string;
}

export interface Flags {
    png:  string;
    svg:  string;
    alt?: string;
}

export interface Idd {
    root:     string;
    suffixes: string[];
}

export interface Maps {
    googleMaps:     string;
    openStreetMaps: string;
}

export interface Name {
    common:     string;
    official:   string;
    nativeName: { [key: string]: Translation };
}

export interface Translation {
    official: string;
    common:   string;
}

export interface PostalCode {
    format: string;
    regex:  string;
}

export enum StartOfWeek {
    Monday = "monday",
}

export enum Status {
    OfficiallyAssigned = "officially-assigned",
    UserAssigned = "user-assigned",
}